requirements/base.txt
# Project's base packages
Django==4.2.4
Markdown==3.4.4
django-environ==0.10.0
django-cors-headers==4.2.0
django-ninja==1.1.0
orjson==3.9.2
pytest==7.4.0
pytest-django==4.5.2
requests==2.31.0
pytest-factoryboy==2.5.1
Faker==19.3.0
django-ninja-jwt==5.2.10
django-ninja-extra==0.20.2
Pillow==10.0.0
ipython==8.12.2
XlsxWriter==3.1.9
openpyxl==3.1.2
celery[redis]==5.3.6
django-celery-beat==2.5.0

# Developer's tools
# Project's base packages
